Added TODO comments to the code
The TODO-s signal what should be done to update to the new version. Most of it is simplification, except for ``@itemprop-inverse`` that means duplicating some code. I have not yet checked whether the duplication can be simplified by factoring out some parts (and it is not sure that this is worth the trouble).